Cursor follows enemies on their turn, enemies can attack and move, small bug fixes
I made it so the cursor follows the enemy unit currently moving so the player can see what is happening.
Enemies can move towards the player: they will spawn a overlap sphere around them in order to see all enemies in range, check which if the closest and move towards it
The enemy can also attack and required me to make some changes to the AOE movement script to account for enemies controlling it.
Made small adjustment to controllable units so the start position which locks them from going out of range updates at the start of a new turn to reflect the unit's new start pos