Adjust how damage dealt, made death state
Adjusted damage script to be on either the ControllableUnit or EnemyUnitAI based on the type of unit, needed to do this as the scriptable object would share the current health amongst all gameobjects that had that class. Also added death states for the enemies and player unit. Finally made a check to see if the currently moving unit destroyed itself and to remove it from the turn order if so