Made slight pause before enemy moves, Unit range now shows when hovering over unit including enemies, commented code, created wait function
Made slight pause before the enemy moves to simulate the idea that they are thinking. Altered the system to made the unit range be showed when hovering over them and also made it so it does the same for enemies. Created the wait function for player to leave their unit in a location without attacking. Commented some code to indicate how to use.