Got healers working
I got the healers working by tagging them and having the system check if aoe is currently a healer. The system will then get the healer information and apply the heal to everything in range. Because of this I needed to make some changes to the system to account for the new tag