Changed the smoke asset, fixed issue with second level
I changed the smoke asset to one that stands out more and is more noticeable. Fixed issue in second level where the wait function would be called twice, was because I forgot to remove the player input system from those tanks, I removed the player input from the prefabs which then applied to all instances of the tanks.