Fixed issue where play can slowly move into wall.
Fixed an issue where if the player kept tapping a direction towards a wall they would be able to enter, add a few lines of code to store a position to send the player too if they stay within a walls trigger. This can be glitched if the player presses a different directional button at the last moment and can get themselves stuck within a object. Currently looking for fix.