Fixed issue where attacks where applied twice
Fixed an issue where attacks applied twice, happened because I added the new hitbox to the tanks, that made the gameobject get stored in the list on in range units twice. Simply created a check to make sure the game object wasn't already in the list when adding a new unit to the units in range.